+++ files/gitlab-cookbooks/package/resources/env_dir.rb
@@ -1,5 +1,5 @@
 #
-# Copyright:: Copyright (c) 2014 GitLab B.V.
+# Copyright:: Copyright (c) 2016 GitLab Inc
 # License:: Apache License, Version 2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,34 +13,39 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-define :env_dir, variables: {}, restarts: [] do
-  env_dir = params[:name]
+resource_name :env_dir
+provides :env_dir
 
-  directory env_dir do
-    recursive true
-  end
+actions :create
+default_action :create
 
-  restarts = params[:restarts]
+property :variables, Hash, default: {}
+property :restarts, Array, default: []
 
-  params[:variables].each do |key, value|
-    file File.join(env_dir, key) do
-      content value
-      restarts.each do |svc|
-        notifies :restart, svc
-      end
-    end
-  end
-
-  if File.directory?(env_dir)
-    deleted_env_vars = Dir.entries(env_dir) - params[:variables].keys - %w(. ..)
+action :create do
+  # Cleaning up non-existent variables
+  if ::File.directory?(name)
+    deleted_env_vars = Dir.entries(name) - variables.keys - %w(. ..)
     deleted_env_vars.each do |deleted_var|
-      file File.join(env_dir, deleted_var) do
+      file ::File.join(name, deleted_var) do
         action :delete
         restarts.each do |svc|
           notifies :restart, svc
         end
+      end
+    end
+  end
+
+  directory name do
+    recursive true
+  end
+
+  variables.each do |key, value|
+    file ::File.join(name, key) do
+      content value
+      restarts.each do |svc|
+        notifies :restart, svc
       end
     end
   end